feat: first attempt to parallelize over the frames
